Move some helpers around.
