+++ spec/helpers/fake_gluster.rb
